feat(Button): Replace deprecated implementation
