+++ packages/react-component-library/src/components/Button/partials/StyledButton.tsx
@@ -2,228 +2,176 @@ import styled, { css } from 'styled-components'
 import { selectors } from '@defencedigital/design-tokens'
 import { rgba } from 'polished'
 
-import {
-  BUTTON_FOCUS_WIDTH,
-  BUTTON_BORDER_RADIUS,
-  BUTTON_VARIANT,
-  BUTTON_COLOR,
-  BUTTON_BORDER_RADIUS_SMALL,
-  BUTTON_SIZE,
-  BUTTON_ICON_POSITION,
-} from '../constants'
-import {
-  ButtonVariantType,
-  ButtonSizeType,
-  ButtonIconPositionType,
-} from '../Button'
-import { StyledText } from './StyledText'
-import { StyledIcon } from './StyledIcon'
+import { BUTTON_ICON_POSITION, BUTTON_VARIANT } from '../constants'
+import { ButtonIconPositionType, ButtonVariantType } from '../Button'
+import { ComponentSizeType, COMPONENT_SIZE } from '../../Forms'
 
 interface StyledButtonProps {
-  $disabled?: boolean
-  $variant?: ButtonVariantType
-  $color?: typeof BUTTON_COLOR.DANGER
-  $size?: ButtonSizeType
-  $iconPosition?: ButtonIconPositionType
+  $variant: ButtonVariantType
+  $size: ComponentSizeType
+  $iconPosition: ButtonIconPositionType
 }
 
-const { color, spacing, fontSize, animation } = selectors
+const { color, spacing, fontSize } = selectors
 
 export const StyledButton = styled.button<StyledButtonProps>`
   position: relative;
+  height: 46px;
   display: inline-flex;
-  flex-direction: row;
+  flex-direction: ${({ $iconPosition }) =>
+    $iconPosition === BUTTON_ICON_POSITION.LEFT ? 'row-reverse' : 'row'};
   align-items: center;
   justify-content: center;
-  border: 1px solid transparent;
-  border-radius: ${BUTTON_BORDER_RADIUS};
+  border-radius: 15px;
   outline: 0;
-  padding: ${spacing('5')} ${spacing('8')} ${spacing('5')} ${spacing('8')};
+  padding: 0 ${spacing('6')};
   font-size: ${fontSize('m')};
-  font-weight: 500;
-  line-height: 1.4;
-  text-align: center;
-  vertical-align: middle;
+  font-weight: 400;
+  text-decoration: none;
   cursor: pointer;
   user-select: none;
-  transition: all ${animation('default')};
-  text-decoration: none;
+  transition: all 75ms cubic-bezier(0, 1.19, 0.82, 0.9);
   white-space: nowrap;
-
-  color: ${color('neutral', 'white')};
-  background-color: ${color('action', '600')};
 
   &:hover {
     text-decoration: none;
-    box-shadow: 0 2px 6px rgba(0, 0, 0, 0.2);
-    background-color: ${color('action', '700')};
   }
 
-  &:focus {
-    border: 1px solid ${color('action', '500')};
-    box-shadow: 0 0 0 ${BUTTON_FOCUS_WIDTH} ${rgba(color('action', '500'), 0.3)};
-  }
-
-  ${({ $iconPosition }) =>
-    $iconPosition === BUTTON_ICON_POSITION.LEFT &&
+  ${({ $size }) =>
+    $size === COMPONENT_SIZE.SMALL &&
     css`
-      flex-direction: row-reverse;
-
-      ${StyledIcon} {
-        margin-left: 0;
-        margin-right: ${spacing('3')};
-      }
+      border-radius: 10px;
+      height: 33px;
+      font-size: ${fontSize('base')};
     `}
 
-  ${({ $disabled }) =>
-    $disabled &&
+  ${({ $variant }) =>
+    $variant === BUTTON_VARIANT.PRIMARY &&
     css`
-      pointer-events: hover;
-      background: ${color('neutral', '000')} !important;
-      color: ${color('neutral', '200')} !important;
-      border: 1px solid ${color('neutral', '100')} !important;
-      cursor: not-allowed;
+      border: 1px solid ${color('action', '600')};
+      color: ${color('neutral', 'white')};
+      background-color: ${color('action', '600')};
+      box-shadow: 0 1px 3px ${rgba(0, 0, 0, 0.1)};
 
+      &:focus,
       &:hover {
-        box-shadow: none;
-        transform: none;
+        background-color: ${color('action', '700')};
+      }
+
+      &:active {
+        background-color: ${color('action', '800')};
       }
     `}
 
-  ${({ $variant, $color }) => {
-    if ($variant === BUTTON_VARIANT.PRIMARY) {
-      return css`
-        color: ${color('neutral', 'white')};
-        background-color: ${color('action', '600')};
+  ${({ $variant }) =>
+    $variant === BUTTON_VARIANT.SECONDARY &&
+    css`
+      color: ${color('neutral', '500')};
+      background-color: ${color('neutral', 'white')};
+      border: 1px solid ${color('neutral', '200')};
+      box-shadow: 0 1px 3px ${rgba(0, 0, 0, 0.1)};
+    `}
 
-        &:hover {
-          background-color: ${color('action', '700')};
-        }
+  ${({ $variant }) =>
+    $variant === BUTTON_VARIANT.TERTIARY &&
+    css`
+      background-color: transparent;
+      background-image: none;
+      border: 1px solid transparent;
+      color: ${color('neutral', '500')};
+      text-decoration: underline;
 
-        ${$color === BUTTON_COLOR.DANGER &&
-        css`
-          background-color: ${color('danger', '700')};
-          color: ${color('neutral', 'white')};
+      &:disabled {
+        text-decoration: none;
+        text-decoration-line: underline;
+        text-decoration-color: transparent;
+      }
 
-          &:hover {
-            background-color: ${color('danger', '800')};
-          }
-
-          &:focus {
-            border-color: ${color('danger', '700')};
-            box-shadow: 0 0 0 ${BUTTON_FOCUS_WIDTH}
-              ${rgba(color('danger', '800'), 0.5)};
-          }
-        `}
-      `
-    }
-
-    if ($variant === BUTTON_VARIANT.SECONDARY) {
-      return css`
+      &:active,
+      &:focus,
+      &:hover {
+        color: ${color('neutral', '500')};
         border: 1px solid ${color('neutral', '200')};
-        color: ${color('neutral', '400')};
+        text-decoration: none;
+        text-decoration-line: underline;
+        text-decoration-color: transparent;
+      }
+    `}
+
+
+  ${({ $variant }) =>
+    ($variant === BUTTON_VARIANT.TERTIARY ||
+      $variant === BUTTON_VARIANT.SECONDARY) &&
+    css`
+      &:focus,
+      &:hover {
         background-color: ${color('neutral', 'white')};
-        background-image: none;
+      }
 
-        &:hover {
-          background-color: ${color('neutral', 'white')};
-          color: ${color('neutral', '600')};
-        }
+      &:active {
+        background-color: ${color('neutral', '100')};
+      }
+    `}
 
-        ${$color === BUTTON_COLOR.DANGER &&
-        css`
-          color: ${color('danger', '700')};
+  ${({ $variant }) =>
+    ($variant === BUTTON_VARIANT.PRIMARY ||
+      $variant === BUTTON_VARIANT.SECONDARY ||
+      $variant === BUTTON_VARIANT.TERTIARY) &&
+    css`
+      &:focus,
+      &:hover {
+        box-shadow: 0 1px 3px ${rgba(0, 0, 0, 0.1)},
+          0 0 0 3px ${color('action', '100')};
+      }
 
-          &:hover {
-            color: ${color('danger', '700')};
-          }
+      &:active {
+        box-shadow: 0 1px 3px transparent, 0 0 0 3px ${color('action', '100')};
+      }
 
-          &:focus {
-            border-color: ${color('danger', '700')};
-            box-shadow: 0 0 0 ${BUTTON_FOCUS_WIDTH}
-              ${rgba(color('danger', '800'), 0.5)};
-          }
-        `}
-      `
-    }
-
-    if ($variant === BUTTON_VARIANT.TERTIARY) {
-      return css`
-        border: 1px solid transparent;
-        color: ${color('neutral', '400')};
-        background-color: transparent;
-        background-image: none;
-
-        &:hover {
-          transform: none;
-          box-shadow: none;
+      &:disabled {
+        &,
+        &:hover,
+        &:active,
+        &:focus {
+          background: ${color('neutral', '000')};
           border: 1px solid ${color('neutral', '200')};
-          background-color: ${color('neutral', 'white')};
-
-          ${StyledText} {
-            &::after {
-              transform: scale(0, 1);
-            }
-          }
+          box-shadow: none;
+          color: ${color('neutral', '300')};
+          cursor: not-allowed;
         }
+      }
+    `}
 
-        ${StyledText} {
-          position: relative;
-          z-index: 0;
+  ${({ $variant }) =>
+    $variant === BUTTON_VARIANT.DANGER &&
+    css`
+      background-color: ${color('danger', '700')};
+      border: 1px solid ${color('danger', '700')};
+      color: ${color('neutral', 'white')};
 
-          &::after {
-            transition: transform ${animation('default')};
-            transform-origin: 0 0;
-            bottom: 2px;
-            left: 0;
-            right: 0;
-            height: 2px;
-            background: ${color('neutral', '200')};
-            content: '';
-            position: absolute;
-            border-radius: 5px;
-            z-index: -1;
-          }
+      &:focus,
+      &:hover {
+        background-color: ${color('danger', '800')};
+        box-shadow: 0 1px 3px ${rgba(0, 0, 0, 0.1)},
+          0 0 0 3px ${color('danger', '100')};
+      }
+
+      &:active {
+        background-color: ${color('danger', '900')};
+        box-shadow: 0 1px 3px transparent, 0 0 0 3px ${color('danger', '100')};
+      }
+
+      &:disabled {
+        &,
+        &:hover,
+        &:active,
+        &:focus {
+          background: ${color('danger', '000')};
+          border: 1px solid ${color('neutral', '200')};
+          box-shadow: none;
+          color: ${color('neutral', '300')};
+          cursor: not-allowed;
         }
-
-        ${$color === BUTTON_COLOR.DANGER &&
-        css`
-          color: ${color('danger', '700')};
-
-          ${StyledText} {
-            &::after {
-              background: ${color('danger', '300')};
-            }
-          }
-        `}
-      `
-    }
-
-    return ``
-  }}
-
-  ${({ $size }) => {
-    if ($size === BUTTON_SIZE.SMALL) {
-      return css`
-        padding: ${spacing('3')} ${spacing('5')};
-        border-radius: ${BUTTON_BORDER_RADIUS_SMALL};
-        font-size: ${fontSize('base')};
-      `
-    }
-
-    // @ts-ignore
-    if ([BUTTON_SIZE.LARGE, BUTTON_SIZE.XLARGE].includes($size)) {
-      return css`
-        padding: ${spacing('6')} ${spacing('9')};
-        font-size: ${fontSize('l')};
-        line-height: 1.45;
-
-        ${StyledText} {
-          transform: translateY(-1px);
-        }
-      `
-    }
-
-    return ``
-  }}
+      }
+    `}
 `
