+++ packages/react-component-library/src/components/Button/partials/StyledIconWrapper.tsx
@@ -1,8 +1,8 @@
 import styled, { css } from 'styled-components'
 import { selectors } from '@defencedigital/design-tokens'
 
-import { BUTTON_E_ICON_POSITION } from '../constants'
-import { ButtonEIconPositionType } from '../ButtonE'
+import { BUTTON_ICON_POSITION } from '../constants'
+import { ButtonIconPositionType } from '../Button'
 import { ComponentSizeType, COMPONENT_SIZE } from '../../Forms'
 
 const { spacing } = selectors
@@ -11,7 +11,7 @@ interface StyledIconWrapperProps {
   $buttonHasText: boolean
   $buttonSize: ComponentSizeType
   $isLoading: boolean
-  $iconPosition: ButtonEIconPositionType
+  $iconPosition: ButtonIconPositionType
 }
 
 export const StyledIconWrapper = styled.span<StyledIconWrapperProps>`
@@ -29,10 +29,10 @@ export const StyledIconWrapper = styled.span<StyledIconWrapperProps>`
   ${({ $buttonHasText, $iconPosition }) =>
     $buttonHasText &&
     css`
-      margin-left: ${$iconPosition === BUTTON_E_ICON_POSITION.RIGHT
+      margin-left: ${$iconPosition === BUTTON_ICON_POSITION.RIGHT
         ? spacing('5')
         : '0'};
-      margin-right: ${$iconPosition === BUTTON_E_ICON_POSITION.LEFT
+      margin-right: ${$iconPosition === BUTTON_ICON_POSITION.LEFT
         ? spacing('5')
         : '0'};
     `};
