--- packages/react-component-library/src/components/ButtonE/constants.ts
@@ -1,13 +1,0 @@
-const BUTTON_E_VARIANT = {
-  PRIMARY: 'primary',
-  SECONDARY: 'secondary',
-  TERTIARY: 'tertiary',
-  DANGER: 'danger',
-} as const
-
-const BUTTON_E_ICON_POSITION = {
-  LEFT: 'left',
-  RIGHT: 'right',
-} as const
-
-export { BUTTON_E_VARIANT, BUTTON_E_ICON_POSITION }
