--- packages/react-component-library/src/components/ButtonE/index.ts
@@ -1,2 +1,0 @@
-export * from './ButtonE'
-export * from './constants'
