--- packages/react-component-library/src/components/ButtonE/partials/StyledText.tsx
@@ -1,9 +1,0 @@
-import styled from 'styled-components'
-
-interface StyledTextProps {
-  $isLoading: boolean
-}
-
-export const StyledText = styled.span<StyledTextProps>`
-  visibility: ${({ $isLoading }) => ($isLoading ? 'hidden' : 'initial')};
-`
