+++ packages/react-component-library/src/components/Timeline/TimelineToolbar.tsx
@@ -6,7 +6,8 @@ import {
   IconZoomOut,
 } from '@defencedigital/icon-library'
 
-import { Button, BUTTON_SIZE, BUTTON_VARIANT } from '../Button'
+import { Button, BUTTON_VARIANT } from '../Button'
+import { COMPONENT_SIZE } from '../Forms'
 import { StyledToolbar } from './partials/StyledToolbar'
 import { StyledToolbarButtons } from './partials/StyledToolbarButtons'
 import { StyledToolbarSeparator } from './partials/StyledToolbarSeparator'
@@ -28,19 +29,21 @@ export const TimelineToolbar: React.FC<TimelineToolbarProps> = ({
       <StyledToolbarButtons>
         <Button
           aria-label="Navigate left"
-          size={BUTTON_SIZE.SMALL}
+          size={COMPONENT_SIZE.SMALL}
           variant={BUTTON_VARIANT.SECONDARY}
           icon={<IconChevronLeft />}
           onClick={movePrevious}
           data-testid="timeline-side-button-left"
+          title="Navigate left"
         />
         <Button
           aria-label="Navigate right"
-          size={BUTTON_SIZE.SMALL}
+          size={COMPONENT_SIZE.SMALL}
           variant={BUTTON_VARIANT.SECONDARY}
           icon={<IconChevronRight />}
           onClick={moveNext}
           data-testid="timeline-side-button-right"
+          title="Navigate right"
         />
       </StyledToolbarButtons>
       {!hideScaling && (
@@ -50,20 +53,22 @@ export const TimelineToolbar: React.FC<TimelineToolbarProps> = ({
             <Button
               aria-label="Zoom out"
               isDisabled={!canZoomOut}
-              size={BUTTON_SIZE.SMALL}
+              size={COMPONENT_SIZE.SMALL}
               variant={BUTTON_VARIANT.SECONDARY}
               icon={<IconZoomOut />}
               onClick={zoomOut}
               data-testid="timeline-toolbar-zoom-out"
+              title="Zoom out"
             />
             <Button
               aria-label="Zoom in"
               isDisabled={!canZoomIn}
-              size={BUTTON_SIZE.SMALL}
+              size={COMPONENT_SIZE.SMALL}
               variant={BUTTON_VARIANT.SECONDARY}
               icon={<IconZoomIn />}
               onClick={zoomIn}
               data-testid="timeline-toolbar-zoom-in"
+              title="Zoom in"
             />
           </StyledToolbarButtons>
         </>
