+++ packages/react-component-library/src/forms/react-hook-form/react-hook-form.stories.tsx
@@ -18,7 +18,7 @@ import { NumberInputE } from '../../components/NumberInputE'
 import { SelectE, SelectEOption } from '../../components/SelectE'
 import { SwitchE, SwitchEOption } from '../../components/SwitchE'
 import { RangeSliderE } from '../../components/RangeSliderE'
-import { ButtonE } from '../../components/ButtonE'
+import { Button } from '../../components/Button'
 import { Fieldset } from '../../components/Fieldset'
 import { sleep } from '../../helpers'
 import { Field } from '../../components/Field'
@@ -269,13 +269,13 @@ export const Example: React.FC<unknown> = () => {
             }}
           />
         </Field>
-        <ButtonE
+        <Button
           type="submit"
           data-testid="form-example-submit"
           isDisabled={isSubmitting}
         >
           Submit
-        </ButtonE>
+        </Button>
       </form>
 
       <pre data-testid="form-example-values">
