+++ packages/react-component-library/src/index.ts
@@ -14,7 +14,6 @@ export * from './components/Avatar'
 export * from './components/Badge'
 export * from './components/Breadcrumbs'
 export * from './components/Button'
-export * from './components/ButtonE'
 export * from './components/ButtonGroup'
 export * from './components/CardFrame'
 export * from './components/Checkbox'
